clio

Struct ClioPath

pub struct ClioPath { /* private fields */ }

A builder for Input and Output.

It is designed to be used to get files related to the one passed in.

e.g. Take an Input of /tmp/foo.svg and have a default Output of /tmp/foo.png

use clio::{Input, Output};

let input = Input::new("/tmp/foo.svg")?;
let mut output_path = input.path().clone();
output_path.set_extension("png");
let output = output_path.create()?;

assert_eq!(output.path().as_os_str().to_string_lossy(), "/tmp/foo.png");

Unlike InputPath and OutputPath it does not validate the path until you try creating/opening it.

However you can add extra validation using the clap parser.

use clap::Parser;
use clio::ClioPath;

#[derive(Parser)]
struct Opt {
    /// path to input file, use '-' for stdin
    #[clap(value_parser)]
    input_path: ClioPath,

    /// path to output file, use '-' for stdout point to directory to use default name
    #[clap(value_parser = clap::value_parser!(ClioPath).default_name("out.bin"))]
    output_file: ClioPath,

    /// path to directory
    #[clap(value_parser = clap::value_parser!(ClioPath).exists().is_dir())]
    log_dir: ClioPath,
}

Implementations

impl ClioPath

pub fn new<S: AsRef<OsStr>>(path: S) -> Result<Self>

Construct a new ClioPath from an string

'-' is treated as stdin/stdout

pub fn std() -> Self

Constructs a new ClioPath of "-" for stdout

pub fn local(path: PathBuf) -> Self

Constructs a new ClioPath for a local path

pub fn set_file_name<S: AsRef<OsStr>>(&mut self, file_name: S)

Updates self.file_name to file_name.

see PathBuf::set_file_name for more details

Examples

use clio::ClioPath;

let mut buf = ClioPath::new("/")?;
assert!(buf.file_name() == None);
buf.set_file_name("bar");
assert!(buf == ClioPath::new("/bar")?);
assert!(buf.file_name().is_some());
buf.set_file_name("baz.txt");
assert!(buf == ClioPath::new("/baz.txt")?);
#[cfg(feature = "http")] {
    let mut p = ClioPath::new("https://example.com/bar.html?x=y#p2")?;
    p.set_file_name("baz.txt");
    assert_eq!(Some("https://example.com/baz.txt?x=y#p2"), p.as_os_str().to_str());  
}

pub fn set_extension<S: AsRef<OsStr>>(&mut self, extension: S) -> bool

Updates self.extension to extension.

see PathBuf::set_extension for more details

Examples

use clio::ClioPath;

let mut p = ClioPath::new("/feel/the")?;

p.set_extension("force");
assert_eq!(ClioPath::new("/feel/the.force")?, p);

p.set_extension("dark_side");
assert_eq!(ClioPath::new("/feel/the.dark_side")?, p);

#[cfg(feature = "http")] {
    let mut p = ClioPath::new("https://example.com/the_force.html?x=y#p2")?;
    p.set_extension("txt");
    assert_eq!(Some("https://example.com/the_force.txt?x=y#p2"), p.as_os_str().to_str());  
}

pub fn add_extension<S: AsRef<OsStr>>(&mut self, extension: S) -> bool

Adds an extension to the end of the self.file_name.

Examples

use clio::ClioPath;

let mut p = ClioPath::new("/tmp/log.txt")?;

p.add_extension("gz");
assert_eq!(ClioPath::new("/tmp/log.txt.gz")?, p);

use clio::ClioPath;

let mut p = ClioPath::new("/tmp/log")?;
p.add_extension("gz");
assert_eq!(ClioPath::new("/tmp/log.gz")?, p);

pub fn push<P: AsRef<Path>>(&mut self, path: P)

Extends self with path.

see PathBuf::push for more details

use clio::ClioPath;

let mut path = ClioPath::new("/tmp")?;
path.push("file.bk");
assert_eq!(path, ClioPath::new("/tmp/file.bk")?);

#[cfg(feature = "http")] {
    let mut p = ClioPath::new("https://example.com/tmp?x=y#p2")?;
    p.push("file.bk");
    assert_eq!(Some("https://example.com/tmp/file.bk?x=y#p2"), p.as_os_str().to_str());
}

pub fn join<P: AsRef<Path>>(&mut self, path: P) -> Self

Creates an owned ClioPath with path adjoined to self.

If path is absolute, it replaces the current path.

See PathBuf::push for more details on what it means to adjoin a path.

Examples

use clio::ClioPath;

assert_eq!(ClioPath::new("/etc")?.join("passwd"), ClioPath::new("/etc/passwd")?);
assert_eq!(ClioPath::new("/etc")?.join("/bin/sh"), ClioPath::new("/bin/sh")?);

#[cfg(feature = "http")] {
    let mut p = ClioPath::new("https://example.com/tmp?x=y#p2")?;
    p.push("file.bk");
    assert_eq!(
        ClioPath::new("https://example.com/tmp?x=y#p2")?.join("file.bk"),
        ClioPath::new("https://example.com/tmp/file.bk?x=y#p2")?);
}

pub fn is_std(&self) -> bool

Returns true if this path is stdin/stout i.e. it was created with -

pub fn is_tty(&self) -> bool

Returns true if this is_std and it would connect to a tty

pub fn is_local(&self) -> bool

Returns true if this path is on the local file system, as opposed to point to stdin/stout or a URL

pub fn ends_with_slash(&self) -> bool

Returns true if this path ends with a /

A trailing slash is often used by command line arguments to refer to a directory that may not exist. e.g. cp foo /tmp/

pub fn files<P>(self, predicate: P) -> Result<Vec<ClioPath>>
where P: FnMut(&ClioPath) -> bool,

If this is a folder returns all the files that match the filter found by looking recursively Otherwise returns just this path

use clio::has_extension;
use clio::ClioPath;

let dir = ClioPath::new("/tmp/foo")?;
for txt_file in dir.files(has_extension("txt"))? {
    txt_file.open()?;
}

pub fn create_with_len(self, size: u64) -> Result<Output>

Create the file with a predetermined length, either using File::set_len or as the content-length header of the http put

pub fn create(self) -> Result<Output>

Create the file at this path and return it as an Output without setting the length

pub fn open(self) -> Result<Input>

Open the file at this path and return it as an Input

pub fn read_all(self) -> Result<CachedInput>

Read the entire the file at this path and return it as an CachedInput

pub fn path(&self) -> &Path

A path represented by this ClioPath If it is - and it is no known if it is in or out then the path will be - If it is - and it is known to be in/out then it will be the pseudo device e.g /dev/stdin If it is a url it will be the path part of the url

use clio::{ClioPath, OutputPath};
use std::path::Path;

let p = ClioPath::new("-")?;

assert_eq!(Path::new("-"), p.path());

let stdout = OutputPath::new("-")?;
let p:&ClioPath = stdout.path();
assert_eq!(Path::new("/dev/stdout"), p.path());

#[cfg(feature = "http")] {
    let p = ClioPath::new("https://example.com/foo/bar.html?x=y#p2")?;

    assert_eq!(Path::new("/foo/bar.html"), p.path());  
}

pub fn as_os_str(&self) -> &OsStr

The original string represented by this ClioPath

pub fn to_os_string(self) -> OsString

Consumes the ClioPath, yielding its internal OsString storage.