clio

Struct Input

pub struct Input { /* private fields */ }

An enum that represents a command line input stream, either Stdin or File

It is designed to be used with the clap crate when taking a file name as an argument to CLI app

use clap::Parser;
use clio::Input;

#[derive(Parser)]
struct Opt {
    /// path to file, use '-' for stdin
    #[clap(value_parser)]
    input_file: Input,
}

Implementations

impl Input

pub fn new<S: TryInto<ClioPath>>(path: S) -> Result<Self>
where Error: From<<S as TryInto<ClioPath>>::Error>,

Constructs a new input either by opening the file or for โ-โ returning stdin

pub fn std() -> Self

Constructs a new input for stdin

pub fn try_from_os_str(path: &OsStr) -> Result<Self, OsString>

Constructs a new input either by opening the file or for โ-โ returning stdin

The error is converted to a OsString so that stuctopt can show it to the user.

It is recommended that you use TryFrom::try_from and clap 3.0 instead.

pub fn len(&self) -> Option<u64>

If input is a file, returns the size of the file, in bytes otherwise if input is stdin returns none.

Examples

let file = clio::Input::new("foo.txt").unwrap();

assert_eq!(Some(3), file.len());

pub fn get_file(&mut self) -> Option<&mut File>

If input is a file, returns a reference to the file, otherwise if input is stdin or a pipe returns none.

pub fn is_empty(&self) -> Option<bool>

Returns a boolean saying if the file is empty, if using stdin returns None

Examples

let file = clio::Input::new("foo.txt").unwrap();

assert_eq!(Some(true), file.is_empty());

pub fn lock<'a>(&'a mut self) -> Box<dyn BufRead + 'a>

If the input is std in locks it, otherwise wraps the file in a buffered reader. This is useful to get the line iterator of the BufRead.

Examples

use std::io::BufRead;
let mut file = clio::Input::new("-")?;

for line in file.lock().lines() {
  println!("line is: {}", line?);
}

pub fn path(&self) -> &ClioPath

Returns the path/url used to create the input

pub fn is_std(&self) -> bool

Returns true if this Input reads from stdin

pub fn is_local(&self) -> bool

Returns true if this Input points to the local file system, as opposed to point to stdin or a URL

pub fn is_tty(&self) -> bool

Returns true if this is stdin and it is connected to a tty

pub fn can_seek(&self) -> bool

Returns true if this Input is a file, and false if this Input is std out or a pipe