clio

Struct InputPath

pub struct InputPath { /* private fields */ }

A builder for Input that validates the path but defers creating it until you call the open method.

It is designed to be used with the clap crate when taking a file name as an argument to CLI app

use clap::Parser;
use clio::InputPath;

#[derive(Parser)]
struct Opt {
    /// path to file, use '-' for stdin
    #[clap(value_parser)]
    input_file: InputPath,
}

Implementations

impl InputPath

pub fn new<S: TryInto<ClioPath>>(path: S) -> Result<Self>
where Error: From<<S as TryInto<ClioPath>>::Error>,

Constructs a new InputPath representing the path and checking that the file exists and is readable

note: even if this passes open may still fail if e.g. the file was delete in between

pub fn std() -> Self

Constructs a new InputPath to stdout (โ-โ)

pub fn is_std(&self) -> bool

Returns true if this InputPath is stdin

pub fn is_tty(&self) -> bool

Returns true if this is stdin and it is connected to a tty

pub fn is_local(&self) -> bool

Returns true if this InputPath is on the local file system, as opposed to point to stdin or a URL

pub fn open(self) -> Result<Input>

Create an Input by opening the file or for โ-โ returning stdin.

This is unlikely to error as the path is checked when the InputPath was created by new but time of use/time of check means that things could have changed in-between e.g. the file could have been deleted.

pub fn path(&self) -> &ClioPath

The original path used to create this InputPath