clio

Struct Output

pub struct Output { /* private fields */ }

A struct that represents a command line output stream, either Stdout or a File along with itโs path

It is designed to be used with the clap crate when taking a file name as an argument to CLI app

use clap::Parser;
use clio::Output;

#[derive(Parser)]
struct Opt {
    /// path to file, use '-' for stdout
    #[clap(value_parser)]
    output_file: Output,

    /// default name for file is user passes in a directory
    #[clap(value_parser = clap::value_parser!(Output).default_name("run.log"))]
    log_file: Output,

    /// Write output atomically using temp file and atomic rename
    #[clap(value_parser = clap::value_parser!(Output).atomic())]
    config_file: Output,
}

Implementations

impl Output

pub fn new<S: TryInto<ClioPath>>(path: S) -> Result<Self>
where Error: From<<S as TryInto<ClioPath>>::Error>,

Constructs a new output either by opening/creating the file or for โ-โ returning stdout

pub fn std() -> Self

Constructs a new output for stdout

pub fn std_err() -> Self

Constructs a new output for stdout

pub fn is_std(&self) -> bool

Returns true if this Output is stout

pub fn is_tty(&self) -> bool

Returns true if this is stdout and it is connected to a tty

pub fn is_local(&self) -> bool

Returns true if this Output is on the local file system, as opposed to point to stdin/stout or a URL

pub fn try_from_os_str(path: &OsStr) -> Result<Self, OsString>

Constructs a new output either by opening/creating the file or for โ-โ returning stdout

The error is converted to a OsString so that stuctopt can show it to the user.

It is recommended that you use TryFrom::try_from and clap 3.0 instead.

pub fn finish(self) -> Result<()>

Syncs the file to disk or closes any HTTP connections and returns any errors or on the file if a regular file For atomic files this must be called to perform the final atomic swap

pub fn lock<'a>(&'a mut self) -> Box<dyn Write + 'a>

If the output is std out locks it. useful in multithreaded context to write lines consistently

Examples

let mut file = clio::Output::new("-")?;

writeln!(file.lock(), "hello world")?;

pub fn get_file(&mut self) -> Option<&mut File>

If output is a file, returns a reference to the file, otherwise if output is stdout or a pipe returns none.

pub fn path(&self) -> &ClioPath

The original path used to create this Output

pub fn can_seek(&self) -> bool

Returns true if this Output is a file, and false if this Output is std out or a pipe