clio

Struct OutputPath

pub struct OutputPath { /* private fields */ }

A builder for Output that validates the path but defers creating it until you call the create method.

The create_with_len allows setting the size before writing. This is mostly useful with the “http” feature for setting the Content-Length header

It is designed to be used with the clap crate when taking a file name as an argument to CLI app

use clap::Parser;
use clio::OutputPath;

#[derive(Parser)]
struct Opt {
    /// path to file, use '-' for stdout
    #[clap(value_parser)]
    output_file: OutputPath,
}

Implementations

impl OutputPath

pub fn new<S: TryInto<ClioPath>>(path: S) -> Result<Self>
where Error: From<<S as TryInto<ClioPath>>::Error>,

Construct a new OutputPath from an string

It checks if an output file could plausibly be created at that path

pub fn std() -> Self

Constructs a new OutputPath of "-" for stdout

pub fn maybe_with_len(self, size: Option<u64>) -> Result<Output>

convert to an normal Output setting the length of the file to size if it is Some

pub fn create_with_len(self, size: u64) -> Result<Output>

Create the file with a predetermined length, either using File::set_len or as the content-length header of the http put

pub fn create(self) -> Result<Output>

Create an Output without setting the length

pub fn path(&self) -> &ClioPath

The original path represented by this OutputPath

pub fn is_std(&self) -> bool

Returns true if this Output is stdout

pub fn is_tty(&self) -> bool

Returns true if this is stdout and it is connected to a tty

pub fn is_local(&self) -> bool

Returns true if this Output is on the local file system, as opposed to point to stout or a URL

pub fn can_seek(&self) -> bool

Returns true if this OutputPath points to a file, and false if this OutputPath is std out or points to a pipe. Note that the file is not opened yet, so there are possible when you open the file it might have changed.